requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-filter==23.3
celery==5.3.4
requests==2.31.0
amqp==5.1.1
pyjwt==2.3.0